Optimized calculation of the DC parts of matrix A and vector b.
See comments in source for details.  For light curves with large numbers
of DC offsets, this calculation was quite slow and the change can
substantially improve the performance.

Further optimization is still possible.

The diagonal block D and the external parameter parts around it are
constant across all periods so could be precomputed, at cost of having
to do a bit more extensive refactoring.  It could also be inverted
ahead of time and the matrix solve operation rewritten taking
advantage of this using something like the Schur complement trick used
in rtdetect.

The calculations of E and the relevant part of R could also be
vectorized, which the compiler doesn't seem to do at the moment.
These are BLAS operations but on my production systems ATLAS isn't
tuned for the specific CPU features so using it doesn't seem to
improve the performance.  A custom routine and a dispatcher using
runtime CPU feature detection is probably a better approach and is
similar to what I'm considering doing for the loop of sin and cos
operations.